Added Data validation of routes and filtering of stocks

+++ api/Controllers/CommentController.cs
@@ -19,14 +19,22 @@ namespace api.Controllers
         [Route("api/comment")]
         public async Task<IActionResult> GetAll()
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }   
             var comments = await _commentRepo.GetAllAsync();
             var commentDto = comments.Select(s => s.ToCommentDto());
             return Ok(commentDto);
         }
 
-        [HttpGet("api/comment/{id}")]
+        [HttpGet("api/comment/{id:int}")]
         public async Task<ActionResult> GetById([FromRoute] int id)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }   
             var comment = await _commentRepo.GetByIdAsync(id);
 
             if (comment == null)
@@ -40,15 +48,23 @@ namespace api.Controllers
         [HttpPost("api/comment")]
         public async Task<IActionResult> Create([FromBody] CreateCommentRequestDto commentdt)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }   
             var commentModel = commentdt.ToCommentFromCreateDTO();
 
             var createdComment = await _commentRepo.CreateAsync(commentModel);
             return CreatedAtAction(nameof(GetById), new { id = createdComment?.Id }, createdComment?.ToCommentDto());
         }
 
-        [HttpPut("api/comment/{id}")]
+        [HttpPut("api/comment/{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }   
             var commentModel = await _commentRepo.UpdateAsync(id, updateDto);
 
             if (commentModel == null)
@@ -59,9 +75,13 @@ namespace api.Controllers
             return Ok(commentModel.ToCommentDto());
         }
 
-        [HttpDelete("api/comment/{id}")]
+        [HttpDelete("api/comment/{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }   
             var deletedComment = await _commentRepo.DeleteAsync(id);
 
             if (deletedComment == null)
